refactor configuration code
