+++ src/plugin/v2/grpc_plugin.ts
@@ -3,7 +3,8 @@ import * as protoLoader from '@grpc/proto-loader';
 import * as promClient from 'prom-client';
 import { GrpcPlugin } from '../../node-plugin';
 import { Logger } from '../../logger';
-import { PluginConfig, SecurityConfig, isAuthTokenValid } from '../../config';
+import { PluginConfig } from './config';
+import { SecurityConfig, isAuthTokenValid } from '../../config/security';
 import { createBrowser, Browser } from '../../browser';
 import { HTTPHeaders, ImageRenderOptions, RenderOptions } from '../../types';
 import {
