+++ tsconfig.json
@@ -7,11 +7,14 @@
     "sourceMap": true,
     "strictNullChecks": true,
     "allowSyntheticDefaultImports": true,
-    "typeRoots": ["node_modules/@types"],
+    "typeRoots": [
+      "node_modules/@types"
+    ],
     "rootDir": "src"
   },
   "include": [
-    "./src/*.ts"
+    "./src/*.ts",
+    "src/config/config.ts"
   ],
   "exclude": [
     "node_modules"
